service/get_add_new_line.go
package pwg

import (
	"context"
	"encoding/xml"
	constants "propwg/common"
	util "propwg/common"
	"propwg/model"
)

// GetAddNewLine receives XML request body, validates it, and responds with XML response.
func (c *Client) GetAddNewLine(ctx context.Context, req *model.GetAddNewLineRequest) (*model.GetAddNewLineResponse, error) {

	req.Session = GenerateAuthSession(c.AuthConfig)

	res, err := c.SendRequest(ctx, req)
	if err != nil {
		c.Logger.Error(err.Error())
		return nil, err
	}

	addNewLineResponse := model.GetAddNewLineResponse{}

	err = xml.Unmarshal([]byte(res), &addNewLineResponse)
	if err != nil {
		c.Logger.Error(err.Error())
		pwgErr := &util.WrappedError{
			Code:    constants.RESPONSE_UNMARSHAL_EXCEPTION_CODE,
			Message: constants.RESPONSE_UNMARSHAL_EXCEPTION_INFO,
			Err:     err,
		}
		return nil, pwgErr
	}
	return &addNewLineResponse, nil
}
